requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
h11==0.14.0
httplib2==0.22.0
idna==3.4
numpy==1.26.0
oauth2==1.9.0.post1
outcome==1.2.0
pandas==2.1.1
pycparser==2.21
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.23.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
wsproto==1.2.0
